Flesh out the aggregation flow
The aggregation flow encompasses the sequence of aggregate requests sent
from the leader to the helper, including the aggregate share request
that is made once the collector requests a batch. At the moment, the
spec does not have enough detail for this flow to be implementable.

This commit aims to fill in the missing pieces, including:

* Aligning VDAF execution with the latest draft, including usage of the
  verification parameters
* Enumerating the conditions under which a report gets rejected
* Handling errors that occur when processing a given report
* Handling invalid state transitions that may occur if the peer does not
  implement the VDAF properly

For the most part, this change meets the requirements for the current
text. However, a couple minor changes are made:

* A message format is defined that allows us to distignuish between the
  first aggregate request and subsequent aggregate requests. Report
  shares will only be sent in the former, resolving an OPEN ISSUE in the
  text.
* The same API endpoint is used for all aggregate request. Previously we
  had specified different URLs for these, which seems needlessly
  complicated from an implementation perspective.

Co-authored-by: David Cook <dcook@letsencrypt.org>
Co-authored-by: Tim Geoghegan <timg@letsencrypt.org>
Co-authored-by: Brandon Pitman <brandon.pitman@gmail.com>